STY: extracted JRO methods
Extracted JRO methods into a methods file.
