+++ pysatMadrigal/instruments/jro_isr.py
@@ -41,8 +41,7 @@ import numpy as np
 from pysat.instruments.methods import general as ps_gen
 from pysat import logger
 
-from pysatMadrigal.instruments.methods import general
-from pysatMadrigal.utils import coords
+from pysatMadrigal.instruments.methods import general, jro
 
 # ----------------------------------------------------------------------------
 # Instrument attributes
@@ -100,15 +99,11 @@ def init(self):
     """Initializes the Instrument object with values specific to JRO ISR
     """
 
-    ackn_str = ' '.join(["The Jicamarca Radio Observatory is a facility of",
-                         "the Instituto Geofisico del Peru operated",
-                         "with support from the NSF AGS-1433968",
-                         "through Cornell University.\n",
-                         general.cedar_rules()])
+    ackn_str = '\n'.join([jro.acknowledgements(), general.cedar_rules()])
 
     logger.info(ackn_str)
     self.acknowledgements = ackn_str
-    self.references = "Depends on the radar experiment; contact PI"
+    self.references = jro.references()
 
     return
 
@@ -340,76 +335,3 @@ def load(fnames, tag=None, inst_id=None, file_type='hdf5'):
     data = data.squeeze(dim=['kindat', 'kinst', 'gdlatr', 'gdlonr'])
 
     return data, meta
-
-
-# ----------------------------------------------------------------------------
-# Local functions
-
-
-def calc_measurement_loc(inst):
-    """ Calculate the instrument measurement location in geographic coordinates
-
-    Parameters
-    ----------
-    inst : pysat.Instrument
-        JRO ISR Instrument object
-
-    Note
-    ----
-    Adds 'gdlat#', 'gdlon#' to the instrument, for all directions that
-    have azimuth and elevation keys that match the format 'eldir#' and 'azdir#'
-
-    """
-
-    az_keys = [kk[5:] for kk in list(inst.data.keys())
-               if kk.find('azdir') == 0]
-    el_keys = [kk[5:] for kk in list(inst.data.keys())
-               if kk.find('eldir') == 0]
-    good_dir = list()
-
-    for i, kk in enumerate(az_keys):
-        if kk in el_keys:
-            try:
-                good_dir.append(int(kk))
-            except ValueError:
-                logger.warning("unknown direction number [{:}]".format(kk))
-
-    # Calculate the geodetic latitude and longitude for each direction
-    if len(good_dir) == 0:
-        raise ValueError("No matching azimuth and elevation data included")
-
-    for dd in good_dir:
-        # Format the direction location keys
-        az_key = 'azdir{:d}'.format(dd)
-        el_key = 'eldir{:d}'.format(dd)
-        lat_key = 'gdlat{:d}'.format(dd)
-        lon_key = 'gdlon{:d}'.format(dd)
-        # JRO is located 520 m above sea level (jro.igp.gob.pe./english/)
-        # Also, altitude has already been calculated
-        gdaltr = np.ones(shape=inst['gdlonr'].shape) * 0.52
-        gdlat, gdlon, _ = coords.local_horizontal_to_global_geo(inst[az_key],
-                                                                inst[el_key],
-                                                                inst['range'],
-                                                                inst['gdlatr'],
-                                                                inst['gdlonr'],
-                                                                gdaltr,
-                                                                geodetic=True)
-
-        # Assigning as data, to ensure that the number of coordinates match
-        # the number of data dimensions
-        inst.data = inst.data.assign({lat_key: gdlat, lon_key: gdlon})
-
-        # Add metadata for the new data values
-        bm_label = "Beam {:d} ".format(dd)
-        inst.meta[lat_key] = {inst.meta.labels.units: 'degrees',
-                              inst.meta.labels.name: bm_label + 'latitude',
-                              inst.meta.labels.desc: bm_label + 'latitude',
-                              inst.meta.labels.min_val: -90.0,
-                              inst.meta.labels.max_val: 90.0,
-                              inst.meta.labels.fill_val: np.nan}
-        inst.meta[lon_key] = {inst.meta.labels.units: 'degrees',
-                              inst.meta.labels.name: bm_label + 'longitude',
-                              inst.meta.labels.desc: bm_label + 'longitude',
-                              inst.meta.labels.fill_val: np.nan}
-
-    return
